requirements.txt
flask
anthropic
python-dotenv
requests
pytesseract
pdfminer.six
python-docx
Pillow